Refactor all functions to utils
