+++ public/loginUser.js
@@ -7,7 +7,6 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-// Login
 const userLogin = (userObj) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const users = yield JSON.parse(localStorage.getItem("users") || "[]");
