+++ public/utils/logout.js
@@ -1,5 +1,1 @@
-const logout = () => {
-    localStorage.removeItem("user");
-    window.location.href = "../";
-};
-export default logout;
+"use strict";
