+++ src/blogs.ts
@@ -1,71 +1,60 @@
-import USERDETAILS from "./models/userDetails";
-import POST from "./models/post";
-import USER from "./models/user";
+import posts from "./data/posts.js";
 
 import { store, retrieve } from "./utils/storeToLocal.js";
 
-import posts from "./data/posts.js";
-import logout from "./utils/logout.js";
+import { USER } from "./models/user.js";
+import { POST } from "./models/post.js";
+import { getUserDetails } from "./utils/userDetails.js";
 
-if (retrieve("user")) {
-  const user: USERDETAILS = retrieve("user", "{}");
-  (document.querySelector(".username") as HTMLElement).innerText =
-    user?.username;
+// Get user details and fill them in at the top if logged in
+// Else redirect to signup page
+if (getUserDetails()) {
+  window.addEventListener("load", async () => {
+    // Check if user is logged in
+    const users: USER[] = await retrieve("users");
+    // // Store the posts
+    let postsArray: POST[] = [];
+    // Check if there are any posts in the storage
+    const storedPosts: POST[] = await retrieve("posts");
 
-  (document.querySelector(".fullname") as HTMLElement).innerText =
-    user?.firstname + user?.lastname;
-} else {
-  window.location.href = "../";
-}
+    if (storedPosts.length === 0) {
+      postsArray = [...posts];
+    } else {
+      postsArray = [...storedPosts];
+    }
 
-(document.querySelector(".logout") as HTMLElement).addEventListener(
-  "click",
-  () => logout()
-);
+    await store("posts", postsArray);
+    const localPosts: POST[] = await retrieve("posts");
 
-window.addEventListener("load", async () => {
-  // Store the posts
-  let postsArray: POST[] = [];
-  const storedPosts: POST[] = await retrieve("posts");
+    // create Elements for posts
+    let postsElement: string = "";
+    for (let i: number = 0; i < localPosts.length - 1; i++) {
+      const preview: string = localPosts[i]?.body.substr(0, 100);
 
-  if (storedPosts.length === 0) {
-    postsArray = [...posts];
-  } else {
-    postsArray = [...storedPosts];
-  }
+      const user: USER | undefined = users.find(
+        (user) => user.id === localPosts[i].userId
+      );
 
-  await store("posts", postsArray);
-  const localPosts: POST[] = await retrieve("posts");
+      postsElement += `
+        <div class="post" id="${posts[i].id}">
+          <h3>${posts[i].title}</h3>
+          <p>${preview}</p>
+          <div class="user">
+            <img src="../public/images/beach.jpg"></img>
 
-  const users: USER[] = await retrieve("users");
-  // create Elements for posts
-  let postsElement: string = "";
-  for (let i: number = 0; i < localPosts.length - 1; i++) {
-    const preview: string = localPosts[i]?.body.substr(0, 100);
+            <p>${
+              user === undefined ? "Anonymous" : user?.username + user?.lastName
+            }</p>
 
-    const user: USER | undefined = users.find(
-      (user) => user.id === localPosts[i].userId
-    );
-
-    postsElement += `
-      <div class="post" id="${posts[i].id}">
-        <h3>${posts[i].title}</h3>
-        <p>${preview}</p>
-        <div class="user">
-          <img src="../public/images/beach.jpg"></img>
-
-          <p>${
-            user === undefined ? "Anonymous" : user?.username + user?.lastName
-          }</p>
-
-          <a id=${posts[i].id}>Read More</a>
+            <a id=${posts[i].id}>Read More</a>
+          </div>
         </div>
-      </div>
-    `;
-  }
+      `;
+    }
 
-  (document.querySelector(".blogs") as HTMLElement).innerHTML += postsElement;
-  document.querySelectorAll(".post").forEach((post) => {
-    post.setAttribute("href", "somelink.php?id" + post.id);
+    (document.querySelector(".blogs") as HTMLElement).innerHTML += postsElement;
+    document.querySelectorAll(".post").forEach((post) => {
+      post.setAttribute("href", "somelink.php?id" + post.id);
+    });
   });
-});
+} else window.location.href = "../";
