+++ src/utils/logout.ts
@@ -1,7 +1,2 @@
-const logout = () => {
-  localStorage.removeItem("user");
 
-  window.location.href = "../";
-};
 
-export default logout;
