+++ tsconfig.json
@@ -5,11 +5,8 @@
     "allowJs": true,
     "outDir": "public",
     "rootDir": "src",
-    // "baseUrl": "./src",
     "strict": true,
     "noImplicitAny": true,
     "esModuleInterop": true
-    // "resolveJsonModule": true
-    // "moduleResolution": "Node"
   }
 }
